+++ src/FileSystem/Exception/HandleFilesException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Solital\Core\FileSystem\Exception;
+
+class HandleFilesException extends \Exception
+{
+}